+++ CHANGES.md
@@ -28,6 +28,7 @@ Change Log
 * Fixed middle mouse button locked glitch [#7137](https://github.com/AnalyticalGraphicsInc/cesium/pull/7137)
 * Fixed an issue where dynamic Entities on terrain would cause a crash in platforms that do not support depth textures such as Internet Explorer [#7103](https://github.com/AnalyticalGraphicsInc/cesium/issues/7103)
 * Fixed an issue where polylines intersecting the near plane would cause blinking. [#6955](https://github.com/AnalyticalGraphicsInc/cesium/pull/6955)
+* Fixed an issue that would cause a crash when removing a post process stage. [#7210](https://github.com/AnalyticalGraphicsInc/cesium/issues/7210)
 
 ### 1.50 - 2018-10-01